pyproject.toml
[project]
name = "dsi324-pdf-extract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fixthaipdf>=0.2.1",
    "fuzzywuzzy>=0.18.0",
    "google>=3.0.0",
    "google-genai>=1.12.1",
    "google-generativeai>=0.8.5",
    "ipykernel>=6.29.5",
    "langchain>=0.3.24",
    "openai>=1.76.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
]
